Refactor + resolving comments
