+++ src/Title.tsx
@@ -1,0 +1,7 @@
+export default function Title() {
+    return (
+        <div>
+            <h1>Todo List</h1>
+        </div>
+    )
+}
